Trying new tests
As an experiment, I give the AI some instructions to do coverage
guided test generation. I gave it explicit instructions to create the
minimum number of tests and prefer property based testing, but it
still wrote tons of tests that duplicate things and barely test what
it should. However, I left them in, because in part this is an
exploration of what AI can do to help.

I'm still not sure if the AI is worse at coding or I'm worse at
telling the AI how to do its job - likely some combination of both.

Some day, I or the AI may get to the point where I can tell it to
consider this test suite and reduce it significantly while still
preserving the integrity of the tests and code.