Get around clang's stupid warning
Maybe that comment is a bit aggressive, but warning about side
effects in a noexcept expression, especially what that expression is
using declval, is just stupid. The whole point of the noexcept
expression is to check if an expression throws, but you can't do that
without a clang warning, even though nobody would ever actually
expect a side effect to take place in that context.